+++ challengerAuto/src/test/java/uk/co/compendiumdev/challenger/restassured/_08_put_update_challenges/C021PutTodosFailUpdateNoTitle422Test.java
@@ -1,5 +1,7 @@
 package uk.co.compendiumdev.challenger.restassured._08_put_update_challenges;
 
+import com.google.gson.Gson;
+import com.google.gson.JsonElement;
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
@@ -11,41 +13,37 @@ import uk.co.compendiumdev.challenger.restassured.api.ChallengesStatus;
 import uk.co.compendiumdev.challenger.restassured.api.RestAssuredBaseTest;
 import uk.co.compendiumdev.challenger.restassured.api.TodosApi;
 
-public class C022PutTodosWrongId400Test extends RestAssuredBaseTest {
+public class C021PutTodosFailUpdateNoTitle422Test extends RestAssuredBaseTest {
 
     @Test
-    void canFailToAmendATodoDueToAttemptToChangeId() {
+    void canFailToAmendATodoDueToMissingTitle() {
 
         Todo amendMe = new TodosApi().getOrCreateAnyExistingTodo();
 
-        int oldId = amendMe.id;
-        int newId = oldId + 1;
-        amendMe.id = newId;
         // amendMe.id = - cannot amend the id as it is auto assigned
         // title is mandatory and must be in the message
+
+        final JsonElement amendTodoJson = new Gson().toJsonTree(amendMe);
+        amendTodoJson.getAsJsonObject().remove("title");
 
         final Response response =
                 RestAssured.given()
                         .header("X-CHALLENGER", xChallenger)
                         .accept("application/json")
                         .contentType("application/json")
-                        .body(amendMe)
-                        .put(apiPath("/todos/" + oldId))
+                        .body(amendTodoJson.toString())
+                        .put(apiPath("/todos/" + amendMe.id))
                         .then()
-                        .statusCode(400)
+                        .statusCode(422)
                         .contentType(ContentType.JSON)
                         .extract()
                         .response();
 
         ErrorMessages error = response.body().as(ErrorMessages.class);
-        Assertions.assertTrue(
-                error.errorMessages
-                        .get(0)
-                        .contains(String.format("Can not amend id from %d to %d", oldId, newId)));
+        Assertions.assertTrue(error.errorMessages.get(0).contains("title : field is mandatory"));
 
         ChallengesStatus statuses = new ChallengesStatus();
         statuses.get();
-        Assertions.assertTrue(
-                statuses.getChallengeNamed("PUT /todos/{id} no amend id (400)").status);
+        Assertions.assertTrue(statuses.getChallengeNamed("PUT /todos/{id} no title (422)").status);
     }
 }