+++ challenger/src/main/resources/content/changes.md
@@ -9,6 +9,10 @@ showads: true
 
 # Change Log
 
+## 2026/07/23 - Expanded Status Code Range
+- Expanded the status codes to cover `431`, `422` and `409`, rather than use `400` for everything
+- Added new challenges for `431` for a Too Long X-Challenger header and `409` mismatched X-Challenger GUIDs
+
 ## 2026/07/17 - API Challenges repo and Swagger UI updates
 
 - Moved to the [API Challenges GitHub repo](https://github.com/eviltester/apichallenges)